Warps (tunnels) now look pretty good
Also, some time ago I assumed 1 LineRenderer per color would be sufficient, and I was wrong since I decided I don't want to draw the lines inside the sphere. So, this includes the switch from LineRenderer to my custom "MultiLineRenderer" which is a wrapper around a list of LineRenderer but provides many of the same properties/methods. That way I didn't need to rewrite much!